libzecale: move some circuit-related code from core into circuits
